# Fix gamma-incorrect interpolation in the 'sharp' OpenGL shader