@kcgen 2022-04-25
Thanks @johnnovak - this is a nice refinement!

At runtime, it seem my GL backend is having trouble compiling the shader (Ubuntu 22.04, OpenGL version 4.60, embedded GPU):

![2022-04-25_09-33](https://user-images.githubusercontent.com/1557255/165133656-fbcc90cf-b8f8-412e-8d81-52d463a36034.png)

Here's just the error text:

``` text
2022-04-25 09:29:55.431 | Error compiling shader: 0:35(20): error: no function with name 'tex2D'
0:35(16): error: no matching function for call to `pow(error, vec4)'; candidates are:
0:35(16): error:    float pow(float, float)
0:35(16): error:    vec2 pow(vec2, vec2)
0:35(16): error:    vec3 pow(vec3, vec3)
0:35(16): error:    vec4 pow(vec4, vec4)
0:36(20): error: no function with name 'tex2D'
0:36(16): error: no matching function for call to `pow(error, vec4)'; candidates are:
0:36(16): error:    float pow(float, float)
0:36(16): error:    vec2 pow(vec2, vec2)
0:36(16): error:    vec3 pow(vec3, vec3)
0:36(16): error:    vec4 pow(vec4, vec4)
0:37(20): error: no function with name 'tex2D'
0:37(16): error: no matching function for call to `pow(error, vec4)'; candidates are:
0:37(16): error:    float pow(float, float)
0:37(16): error:    vec2 pow(vec2, vec2)
0:37(16): error:    vec3 pow(vec3, vec3)
0:37(16): error:    vec4 pow(vec4, vec4)
0:38(20): error: no function with name 'tex2D'
0:38(16): error: no matching function for call to `pow(error, vec4)'; candidates are:
0:38(16): error:    float pow(float, float)
0:38(16): error:    vec2 pow(vec2, vec2)
0:38(16): error:    vec3 pow(vec3, vec3)
0:38(16): error:    vec4 pow(vec4, vec4)
0:42(12): warning: `c_s0t0' used uninitialized
0:42(22): warning: `c_s1t0' used uninitialized
0:42(31): warning: `c_s0t0' used uninitialized
0:43(12): warning: `c_s0t1' used uninitialized
0:43(22): warning: `c_s1t1' used uninitialized
0:43(31): warning: `c_s0t1' used uninitialized
```