@johnnovak 2022-04-26
Pushed up a revised version which should be much more robust in dealing with different OpenGL versions. Let me know if you're still experiencing any issues.

Btw, just for the heck of it I checked the other built-in shaders and no surprise, they're all gamma incorrect except for the the CRT ones.

The artefacts are quite noticeable on `advinterp2x`, `advinterp3x`, `advmame2x`, `advmame3x` too, as shown below. I'm inclined to apply the same gamma fixes to these as well, what do you think? They're already doing their own interpolation anyway, so it's just a matter of adding the `pow` stuff on the way in & out.

I don't care about the `scan*`, `rgb*` and `tv*` ones at all; they're really just some ancient poor man's CRT emulation attempts and generally result in a horrible aliased mess (I'd personally just remove them altogether).

Just a heads up, I will be away for the next two days, so I won't reply to any comments until about Friday.

## `advmame3x` gamma-incorrect built-in version

![advmame3x-wrong](https://user-images.githubusercontent.com/698770/165221481-7593c48d-7a9e-4de7-acb1-d9e9ad5b76f8.png)

## `advmame3x` gamma-corrected

![advmame3x-correct](https://user-images.githubusercontent.com/698770/165221470-99da9b8a-b892-469b-9c67-28d39e3bb541.png)