@johnnovak 2022-04-26
And here's my gamma test image in case you need it (GitHub doesn't let me attach BMP files, so you'll need to convert it if you want to display it with QPEG).

![gamma](https://user-images.githubusercontent.com/698770/165222595-8db3b834-1fcc-4d36-b91d-6b0f9e80409f.png)