@kcgen 2022-04-26
> Pushed up a revised version which should be much more robust in dealing with different OpenGL versions. 

Compiles and run without issue now; thanks!

I ran a side-by-side speed comparison (quake demo), `main` runs at 200+ FPS fullscreen (4k display) and this branch in the 120s. 

```ini
[sdl]
fullscreen = true
output = opengl
vsync = false

[render]
scaler = none
aspect = true
glshader = sharp

[cpu]
core = dynamic
cycles = max

[dosbox]
startup_verbosity = quiet

[dos]
xms=true

[mixer]
nosound = true

[joystick]
joysticktype = disabled

[autoexec]
mount c .
c:
dosbench.bat
exit
```

And in windowed mode, the performance drop is even more dramatic: `main` runs at ~160 to 170 FPS, this branch in the 50s to 60s FPS, changing only the following:

```ini
fullscreen = false
output = opengl
vsync = true
windowresolution = 3840x2160
```

Are there any cheaper approximations / lookup tables / etc? (maybe for 256 colors, a pow table is possible?)