@kcgen 2022-04-26
> I'm inclined to apply the same gamma fixes to these as well, what do you think

:+1: