@johnnovak 2022-04-26
Great to hear that it works for you now.

No, there isn't much to do to speed it up (apart from not doing the interpolation correctly in the first place, exactly what the previous version was doing). You know the saying --- the fastest car is the one without a brake. Speed is not always everything. Another one I like: "I could compute incorrect results for you much faster". See, this is what graphics programming is all about; all sorts of speed hacks you can get away with with minimal artifacts. It's always a tradeoff between doing things correctly or faster (sometimes *a lot* faster).

LUT tables are definitely out of the question here to gain performance on a GPU, even on regular CPUs they have been rendered ineffective at least 15-20 years ago (they almost always make things a lot slower, at least for for arithmetic function style lookups).

I think the "correct" way to look at it is that the proper version gives you quite good performance, and of course the speedhacked incorrect version is a lot faster (not doing calcuations is always going to be faster). I'm inclined to make the fixed version the default, but also include the old one as a "sharp-fast" builtin. That way we would default to good fidelity out of the box with no artefacts, and I think it doesn't actually matter that we're doing "only" say 200 vs 120 FPS, as long as it's well above the monitor refresh rate. Then people can use the fast version on old hardware or if the situation demands it (e.g. your case of playing Quake on 4K in windowed mode, you might argue that's more of a special case than the norm).

Out of curiosity, what DOS resolution did you use? 320x200 or something higher?