@GranMinigun 2022-04-26
Hey John,

Thanks for raising the problem and providing solution. We are well aware that it's always about trade-offs, and primary concerns here are due to desire to support Raspberry Pi 4, and also due to the fact that higher resolution and high refresh rate displays will inevitably become the norm, given time (1440p 144 Hz becomes default in PC space, consoles push for 4K output, and there're also Retina displays in Apple devices).

Gamma correction itself doesn't impose a major performance hit, it's the custom sample interpolation specifically that hits real hard. I don't expect that everyone does interpolation differently (and sRGB textures require extension anyway), so how about doing just gamma correction and testing for some time with different vendors and drivers to see if it looks correctly everywhere?