@Grounded0 2022-04-26
Here's Quake timedemo running at 640x480 on 15" MacBook Pro (2017) with system maxed out. There's actually less performance leeway available than people would think:
<img width="1312" alt="Screenshot 2022-04-26 at 15 37 54" src="https://user-images.githubusercontent.com/48192273/165301719-7f9c4780-d4ef-4b30-b1fd-f64c5675d0b6.png">